found the bugs. Need solution to solve it.
There is a bug in the equation to find the dA final layer that makes the cost more after each epoch. 
This bug is preventing the dZ to be wrong hence making the whole model go crazy stupid.

The second problem is with the learning rate decay which after a while is going back to the previous thing.